requirements.txt
torch>=2.0
torchvision>=0.15
onnxruntime-directml>=1.15
numpy>=1.21
Pillow>=9.0
scikit-learn>=1.0
opencv-python>=4.5
